backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
rembg
onnxruntime
aiomysql
google-generativeai
pydantic
httpx
python-dotenv
cos-python-sdk-v5
